+++ docs/changelog.rst
@@ -9,6 +9,7 @@ CHANGELOG
 
 * Ensure trek duration is a positive number
 * Fix cirkwi exports (second try)
+* Fix public PDF templates
 
 
 0.33.3 (2015-04-01)